+++ microservice-transaction-sample/web/src/app/_api/authentication-util.ts
