+++ microservice-transaction-sample/web/src/app/_api/create-post.ts
@@ -1,13 +1,11 @@
 import { $ } from "zx";
-import { getUserId } from "./authenticationUtil";
+import { getUserId } from "./authentication-util";
 
 export const createPost = async (
   message: string,
   serverId: number
 ): Promise<void> => {
   const userId = await getUserId();
-  // TODO: remove postId from parameters. Auto-generate postId.
-  const postId = 1;
   let commandName = "";
   switch (serverId) {
     case 1:
@@ -17,7 +15,7 @@ export const createPost = async (
       commandName = "CreatePostOnCassandra";
   }
   const exec =
-    await $`cd ..; ./gradlew :client:run --args="${commandName} ${postId} ${userId} ${message}"`;
+    await $`cd ..; ./gradlew :client:run --args='${commandName} ${userId} "${message}"'`;
   if (exec.exitCode !== 0) {
     throw new Error(exec.stderr);
   }
