+++ microservice-transaction-sample/web/src/app/user/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from "next/server";
-import { getUser } from "../_api/getUser";
+import { getUser } from "../_api/get-user";
 
 export const GET = async (request: Request) => {
   const response = await getUser(1);
